Timeouts are 64 bits

+++ src/context.c
@@ -1992,7 +1992,7 @@ getdns_extension_set_eventloop(struct getdns_context* context,
 
 getdns_return_t
 getdns_context_schedule_timeout(struct getdns_context* context,
-    getdns_transaction_t id, uint16_t timeout, getdns_timeout_callback callback,
+    getdns_transaction_t id, uint64_t timeout, getdns_timeout_callback callback,
     void* userarg) {
     RETURN_IF_NULL(context, GETDNS_RETURN_INVALID_PARAMETER);
     RETURN_IF_NULL(callback, GETDNS_RETURN_INVALID_PARAMETER);
